backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
openai==0.27.0
PyJWT==2.3.0
gitpython==3.1.24
python-jose==3.3.0
requests==2.28.1
pillow==9.3.0
python-magic==0.4.27
boto3==1.26.137
google-auth==2.16.0
google-auth-oauthlib==0.8.0
authlib==1.2.0
